# Optimise polygon merging in complex polygons