Think I managed to make a tiny perf boost to complex polygons... and i think I managed to same a little bit of memory in the process.

 Method |       Mean |    StdDev | Scaled | Scaled-StdDev |     Gen 0 |    Gen 1 |    Gen 2 | Allocated |
---------------------------- |----------- |---------- |------- |-------------- |---------- |--------- |--------- |---------- |
  'Complex Polygon - Master' | 11.6810 ms | 0.0674 ms |   1.00 |          0.00 | 9422.9167 | 672.9167 | 672.9167 |  35.13 MB |
 'Complex Polygon - Updated' |  6.6927 ms | 0.2831 ms |   0.57 |          0.02 |  706.2500 | 677.0833 | 677.0833 |   7.63 MB |`